-Finished setting up general layout
