+++ app/(tabs)/_layout.tsx
@@ -12,9 +12,9 @@ export default function TabLayout() {
         }}
       />
       <Tabs.Screen
-        name="settings"
+        name="garden"
         options={{
-          title: 'Settings',
+          title: 'Garden',
           tabBarIcon: ({ color }) => <FontAwesome size={28} name="cog" color={color} />,
         }}
       />
