--- app/(tabs)/garden.tsx
@@ -1,17 +1,0 @@
-import { View, Text, StyleSheet } from 'react-native';
-
-export default function Tab() {
-  return (
-    <View style={styles.container}>
-      <Text>Tab</Text>
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-});
