+++ app/(tabs)/garden/_layout.tsx
@@ -1,0 +1,7 @@
+import { Stack } from 'expo-router';
+
+export default function GardenLayout() {
+  return (
+    <Stack screenOptions={{headerShown:false}}></Stack>
+  );
+}
